+++ UVtools.Core/Operations/OperationRaftRelief.cs
@@ -10,8 +10,12 @@ using Emgu.CV;
 using Emgu.CV.CvEnum;
 using Emgu.CV.Structure;
 using System;
+using System.ComponentModel;
 using System.Drawing;
+using System.Text;
 using System.Threading.Tasks;
+using Emgu.CV.Util;
+using UVtools.Core.EmguCV;
 using UVtools.Core.Extensions;
 using UVtools.Core.FileFormats;
 
@@ -23,9 +27,17 @@ public class OperationRaftRelief : Operation
     #region Enums
     public enum RaftReliefTypes : byte
     {
+        [Description("Relief: Drill raft to relief pressure and remove some mass")]
         Relief,
+
+        [Description("Dimming: Darkens the raft to cure it less")]
         Dimming,
-        Decimate
+
+        [Description("Decimate: Remove the raft and keep supports only on the plate")]
+        Decimate,
+
+        [Description("Tabs: Creates tabs around the raft to easily insert a tool under it and detach the raft from build plate")]
+        Tabs
     }
     #endregion
         
@@ -38,6 +50,9 @@ public class OperationRaftRelief : Operation
     private byte _wallMargin = 40;      // +/- 2mm
     private byte _holeDiameter = 80;    // +/- 4mm
     private byte _holeSpacing = 40;     // +/- 2mm
+    private byte _tabBrightness = byte.MaxValue;
+    private ushort _tabTriangleBase = 200;
+    private ushort _tabTriangleHeight = 250;
 
     #endregion
 
@@ -57,7 +72,20 @@ public class OperationRaftRelief : Operation
 
     public override Enumerations.LayerRangeSelection StartLayerRangeSelection =>
         Enumerations.LayerRangeSelection.None;
-        
+
+    public override string? ValidateInternally()
+    {
+        var sb = new StringBuilder();
+        if (_reliefType == RaftReliefTypes.Tabs)
+        {
+            if(_tabTriangleHeight == 0) sb.AppendLine("The tab height can't be 0");
+            if(_tabTriangleBase == 0) sb.AppendLine("The tab base can't be 0");
+            if(_tabBrightness == 0) sb.AppendLine("The tab brightness can't be 0");
+        }
+
+        return sb.ToString();
+    }
+
     public override string ToString()
     {
         var result = $"[{_reliefType}] [Mask layer: {_maskLayerIndex}] [Ignore: {_ignoreFirstLayers}] [B: {_brightness}] [Dilate: {_dilateIterations}] [Wall margin: {_wallMargin}] [Hole diameter: {_holeDiameter}] [Hole spacing: {_holeSpacing}]";
@@ -75,8 +103,6 @@ public class OperationRaftRelief : Operation
     #endregion
 
     #region Properties
-    public static Array RaftReliefItems => Enum.GetValues(typeof(RaftReliefTypes));
-
     public RaftReliefTypes ReliefType
     {
         get => _reliefType;
@@ -86,12 +112,15 @@ public class OperationRaftRelief : Operation
             RaisePropertyChanged(nameof(IsRelief));
             RaisePropertyChanged(nameof(IsDimming));
             RaisePropertyChanged(nameof(IsDecimate));
+            RaisePropertyChanged(nameof(IsTabs));
+            RaisePropertyChanged(nameof(BrightnessPercent));
         }
     }
 
     public bool IsRelief => _reliefType == RaftReliefTypes.Relief;
     public bool IsDimming => _reliefType == RaftReliefTypes.Dimming;
     public bool IsDecimate => _reliefType == RaftReliefTypes.Decimate;
+    public bool IsTabs => _reliefType == RaftReliefTypes.Tabs;
 
     public uint MaskLayerIndex
     {
@@ -115,7 +144,7 @@ public class OperationRaftRelief : Operation
         }
     }
 
-    public decimal BrightnessPercent => Math.Round(_brightness * 100 / 255M, 2);
+    public decimal BrightnessPercent => Math.Round((_reliefType == RaftReliefTypes.Tabs ? _tabBrightness : _brightness) * 100 / 255M, 2);
 
     public byte DilateIterations
     {
@@ -140,13 +169,36 @@ public class OperationRaftRelief : Operation
         get => _holeSpacing;
         set => RaiseAndSetIfChanged(ref _holeSpacing, value);
     }
+
+    public byte TabBrightness
+    {
+        get => _tabBrightness;
+        set => RaiseAndSetIfChanged(ref _tabBrightness, Math.Max((byte)1, value));
+    }
+
+    public ushort TabTriangleBase
+    {
+        get => _tabTriangleBase;
+        set => RaiseAndSetIfChanged(ref _tabTriangleBase, Math.Max((ushort)5, value));
+    }
+
+    public ushort TabTriangleHeight
+    {
+        get => _tabTriangleHeight;
+        set
+        {
+            if (!RaiseAndSetIfChanged(ref _tabTriangleHeight, Math.Max((ushort)5, value))) return;
+            RaisePropertyChanged(nameof(BrightnessPercent));
+        } 
+    }
+
     #endregion
 
     #region Equality
 
     protected bool Equals(OperationRaftRelief other)
     {
-        return _reliefType == other._reliefType && _maskLayerIndex == other._maskLayerIndex && _ignoreFirstLayers == other._ignoreFirstLayers && _brightness == other._brightness && _dilateIterations == other._dilateIterations && _wallMargin == other._wallMargin && _holeDiameter == other._holeDiameter && _holeSpacing == other._holeSpacing;
+        return _reliefType == other._reliefType && _maskLayerIndex == other._maskLayerIndex && _ignoreFirstLayers == other._ignoreFirstLayers && _brightness == other._brightness && _dilateIterations == other._dilateIterations && _wallMargin == other._wallMargin && _holeDiameter == other._holeDiameter && _holeSpacing == other._holeSpacing && _tabTriangleBase == other._tabTriangleBase && _tabTriangleHeight == other._tabTriangleHeight;
     }
 
     public override bool Equals(object? obj)
@@ -159,7 +211,18 @@ public class OperationRaftRelief : Operation
 
     public override int GetHashCode()
     {
-        return HashCode.Combine((int) _reliefType, _maskLayerIndex, _ignoreFirstLayers, _brightness, _dilateIterations, _wallMargin, _holeDiameter, _holeSpacing);
+        var hashCode = new HashCode();
+        hashCode.Add((int) _reliefType);
+        hashCode.Add(_maskLayerIndex);
+        hashCode.Add(_ignoreFirstLayers);
+        hashCode.Add(_brightness);
+        hashCode.Add(_dilateIterations);
+        hashCode.Add(_wallMargin);
+        hashCode.Add(_holeDiameter);
+        hashCode.Add(_holeSpacing);
+        hashCode.Add(_tabTriangleBase);
+        hashCode.Add(_tabTriangleHeight);
+        return hashCode.ToHashCode();
     }
 
     #endregion
@@ -176,7 +239,6 @@ public class OperationRaftRelief : Operation
         var anchor = new Point(-1, -1);
         var kernel = EmguExtensions.Kernel3x3Rectangle;
 
-
         uint firstSupportLayerIndex = _maskLayerIndex;
         if (firstSupportLayerIndex <= 0)
         {
@@ -185,13 +247,13 @@ public class OperationRaftRelief : Operation
             for (; firstSupportLayerIndex < layerCount; firstSupportLayerIndex++)
             {
                 progress.ThrowIfCancellationRequested();
-                progress++;
                 supportsMat = GetRoiOrDefault(SlicerFile[firstSupportLayerIndex].LayerMat);
                 var circles = CvInvoke.HoughCircles(supportsMat, HoughModes.Gradient, 1, 5, 80, 35, 5, 200);
                 if (circles.Length >= minSupportsRequired) break;
                     
                 supportsMat.Dispose();
                 supportsMat = null;
+                progress++;
             }
         }
         else
@@ -265,6 +327,109 @@ public class OperationRaftRelief : Operation
                 case RaftReliefTypes.Decimate:
                     supportsMat.CopyTo(target);
                     break;
+                case RaftReliefTypes.Tabs:
+                {
+                    using var contours = mat.FindContours(RetrType.External);
+                    var span = mat.GetDataByteSpan();
+
+                    var minX = 10;
+                    var maxX = mat.Size.Width - 10;
+                    var minY = 10;
+                    var maxY = mat.Size.Height - 10;
+
+                    var triangleBaseRadius = _tabTriangleBase / 2;
+                    var triangleHeightRadius = _tabTriangleHeight / 2;
+
+                    var directions = new[]
+                    {
+                        new Point(0, -1),  // Up
+                        new Point(1, 0),   // Right
+                        new Point(0, 1),   // Down
+                        new Point(-1, 0),  // Left
+                    };
+
+                    var color = new MCvScalar(_tabBrightness);
+
+                    for (var i = 0; i < contours.Size; i++)
+                    {
+                        using var contour = new EmguContour(contours[i]);
+                        if(contour.Centroid.IsAnyNegative() || contour.Area < 10000) continue;
+
+                        for (var dir = 0; dir < directions.Length; dir++)
+                        {
+                            var direction = directions[dir];
+
+                            var foundFirstWhite = false;
+                            var foundPoint = false;
+
+                            var x = contour.Centroid.X;
+                            var y = contour.Centroid.Y;
+
+                            while (!foundPoint 
+                                   && x >= minX && x <= maxX && y >= minY && y <= maxY
+                                   && contour.Bounds.Contains(x, y))
+                            {
+                                var pixel = span[mat.GetPixelPos(x, y)];
+
+                                if (pixel > 0)
+                                {
+                                    if (!foundFirstWhite)
+                                    {
+                                        foundFirstWhite = true;
+                                        continue;
+                                    }
+
+                                    if (CvInvoke.PointPolygonTest(contours[i], new PointF(x, y), false) == 0) // Must be on edge
+                                    {
+                                        foundPoint = true;
+                                        break;
+                                    }
+                                }
+                                
+
+                                x += direction.X;
+                                y += direction.Y;
+                            }
+
+                            if(!foundPoint) continue;
+
+                            var polygon = new Point[3];
+
+                            switch (dir)
+                            {
+                                case 0: // Up
+                                    polygon[0] = new Point(Math.Max(10, x - triangleBaseRadius), y);
+                                    polygon[1] = new Point(Math.Min(mat.Width - 10, x + triangleBaseRadius), y);
+                                    polygon[2] = new Point(x, Math.Max(10, y - triangleHeightRadius));
+                                    break;
+                                case 1: // Right
+                                    polygon[0] = new Point(x, Math.Max(10, y - triangleBaseRadius));
+                                    polygon[1] = new Point(x, Math.Min(mat.Width - 10, y + triangleBaseRadius));
+                                    polygon[2] = new Point(Math.Min(mat.Width - 10, x + triangleHeightRadius), y);
+                                    break;
+                                case 2: // Down
+                                    polygon[0] = new Point(Math.Max(10, x - triangleBaseRadius), y);
+                                    polygon[1] = new Point(Math.Min(mat.Width - 10, x + triangleBaseRadius), y);
+                                    polygon[2] = new Point(x, Math.Min(mat.Height - 10, y + triangleHeightRadius));
+                                    break;
+                                case 3: // Left
+                                    polygon[0] = new Point(x, Math.Max(10, y - triangleBaseRadius));
+                                    polygon[1] = new Point(x, Math.Min(mat.Width - 10, y + triangleBaseRadius));
+                                    polygon[2] = new Point(Math.Max(10, x - triangleHeightRadius), y);
+                                    break;
+                            }
+
+                            CvInvoke.Ellipse(mat, new Point(x, y), new Size(triangleBaseRadius, (int)(triangleBaseRadius / 1.5)), 90 * dir, 0, 180, EmguExtensions.WhiteColor, -1, LineType.AntiAlias);
+                            using var vec = new VectorOfPoint(polygon);
+                            CvInvoke.FillPoly(mat, vec, color, LineType.AntiAlias);
+                        }
+                        
+                    }
+
+                    break;
+                }
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(ReliefType));
             }
 
             ApplyMask(original, target);