+++ UVtools.Core/SystemOS/SystemAware.cs
@@ -271,4 +271,25 @@ public static class SystemAware
         return stringBuilder.ToString();
     }
 
+    /// <summary>
+    /// Gets if is running under Linux and under AppImage format
+    /// </summary>
+    public static bool IsRunningLinuxAppImage(out string? path)
+    {
+        path = null;
+        if (!OperatingSystem.IsLinux()) return false;
+        path = Environment.GetEnvironmentVariable("APPIMAGE");
+        return !string.IsNullOrWhiteSpace(path);
+    }
+
+    /// <summary>
+    /// Gets if is running under Linux and under AppImage format
+    /// </summary>
+    /// <returns></returns>
+    public static bool IsRunningLinuxAppImage() => IsRunningLinuxAppImage(out _);
+
+    /// <summary>
+    /// Gets if is running under MacOS and under app format
+    /// </summary>
+    public static bool IsRunningMacOSApp => OperatingSystem.IsMacOS() && AppContext.BaseDirectory.EndsWith(".app/Contents/MacOS");
 }