+++ UVtools.WPF/Program.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Diagnostics;
 using System.Globalization;
+using System.IO;
 using System.Runtime.ExceptionServices;
 using Avalonia;
 using Projektanker.Icons.Avalonia;